Protect against buzz race condition
In situations where there are multiple buzzes very close to each other
in time, the initial game states sent back from server may not include
the player's buzz yet. In this situation, we should not interpret the
lack of player buzz info as "host has reset player's buzzer" and reset
the buzzer permanently - we should revert that reset as soon as we find
out the player's buzz was recorded and sent down in state.