Better fix for race condition
In previous commit 47ab534, there would be a visible flash, when the
buzzer goes from "buzzed" to "buzz" to "buzzed" again (as the game
state info is sent down in succession). Instead, it will be cleaner to
delay the "reset" by 500 ms, which should both allow buzzer resets by
host, and also race condition support as well.